requirements.txt
django==3.1.6
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
python-dotenv==0.15.0
requests==2.25.1
psycopg2==2.8.6
gunicorn==20.0.4
django-heroku==0.3.1